+++ pyvela/pyvela/pyvela_script.py
@@ -1,10 +1,6 @@
-import datetime
-import getpass
 import json
 import os
-import platform
 import shutil
-import sys
 from argparse import ArgumentParser, ArgumentDefaultsHelpFormatter
 from copy import deepcopy
 from warnings import warn
@@ -14,53 +10,7 @@ import numpy as np
 import pint
 import pint.models
 
-import pyvela
 from pyvela import SPNTA
-from pyvela import Vela as vl
-
-
-def info_dict(args):
-    info_dict = {
-        "input": {
-            "par_file": (
-                os.path.basename(args.par_file) if args.par_file is not None else None
-            ),
-            "tim_file": (
-                os.path.basename(args.tim_file) if args.tim_file is not None else None
-            ),
-            "jlso_file": (
-                os.path.basename(args.jlso_file) if args.jlso_file is not None else None
-            ),
-            "prior_file": (
-                os.path.basename(args.prior_file)
-                if args.prior_file is not None
-                else None
-            ),
-            "cheat_prior_scale": args.cheat_prior_scale,
-        },
-        "sampler": {
-            "nsteps": args.nsteps,
-            "burnin": args.burnin,
-            "thin": args.thin,
-        },
-        "env": {
-            "launch_time": datetime.datetime.now().isoformat(),
-            "user": getpass.getuser(),
-            "host": platform.node(),
-            "os": platform.platform(),
-            "julia_threads": vl.nthreads(),
-            "python": sys.version,
-            "julia": str(vl.VERSION),
-            "pyvela": pyvela.__version__,
-            "pint": pint.__version__,
-            "emcee": emcee.__version__,
-        },
-    }
-
-    if args.truth is not None:
-        info_dict["input"]["truth_par_file"] = os.path.basename(args.truth)
-
-    return info_dict
 
 
 def parse_args(argv):
@@ -199,94 +149,8 @@ def prepare_outdir(args):
     if args.prior_file is not None:
         shutil.copy(args.prior_file, args.outdir)
 
-    summary_info = info_dict(args)
-    with open(f"{args.outdir}/summary.json", "w") as summary_file:
-        json.dump(summary_info, summary_file, indent=4)
-
     if args.truth is not None:
         shutil.copy(args.truth, args.outdir)
-
-
-def get_true_values(spnta: SPNTA, args):
-    true_model = pint.models.get_model(args.truth)
-    return (
-        np.array(
-            [
-                (
-                    (true_model[par].value if par != "F0" else 0.0)
-                    if par in true_model
-                    else np.nan
-                )
-                for par in spnta.param_names
-            ]
-        )
-        * spnta.scale_factors
-    )
-
-
-def save_spnta_attrs(spnta: SPNTA, args):
-    np.savetxt(f"{args.outdir}/param_names.txt", spnta.param_names, fmt="%s")
-    np.savetxt(f"{args.outdir}/param_prefixes.txt", spnta.param_prefixes, fmt="%s")
-    np.savetxt(f"{args.outdir}/param_units.txt", spnta.param_units, fmt="%s")
-    np.savetxt(f"{args.outdir}/param_scale_factors.txt", spnta.scale_factors)
-
-    if args.truth is not None:
-        np.savetxt(f"{args.outdir}/param_true_values.txt", get_true_values(spnta, args))
-
-
-def save_new_parfile(
-    spnta: SPNTA, params: np.ndarray, param_uncertainties: np.ndarray, filename: str
-):
-    param_vals = spnta.rescale_samples(params)
-    param_errs = spnta.rescale_samples(param_uncertainties)
-
-    model1 = (
-        deepcopy(spnta.model_pint_modified)
-        if spnta.model_pint_modified is not None
-        else spnta.model_pint
-    )
-    for pname, pval, perr in zip(spnta.param_names, param_vals, param_errs):
-        if pname in model1:
-            model1[pname].value = (
-                pval
-                if pname != "F0"
-                else (
-                    np.longdouble(spnta.model.param_handler._default_params_tuple.F_.x)
-                    + pval
-                )
-            )
-            model1[pname].uncertainty_value = perr
-        else:
-            warn(f"Parameter {pname} not found in the PINT TimingModel!")
-
-    model1.write_parfile(filename)
-
-
-def save_resids(spnta: SPNTA, params: np.ndarray, outdir: str) -> None:
-    wb = spnta.wideband
-
-    ntoas = len(spnta.toas)
-    mjds = spnta.mjds
-    tres = spnta.time_residuals(params)
-    tres_w = spnta.whitened_time_residuals(params)
-    terr = spnta.scaled_toa_unceritainties(params)
-
-    res_arr = np.zeros((ntoas, 3 * (1 + int(wb)) + 1))
-    res_arr[:, 0] = mjds
-    res_arr[:, 1] = tres
-    res_arr[:, 2] = tres_w
-    res_arr[:, 3] = terr
-
-    if wb:
-        dres = spnta.dm_residuals(params)
-        dres_w = spnta.whitened_dm_residuals(params)
-        derr = spnta.scaled_dm_unceritainties(params)
-
-        res_arr[:, 4] = dres
-        res_arr[:, 5] = dres_w
-        res_arr[:, 6] = derr
-
-    np.savetxt(f"{outdir}/residuals.txt", res_arr)
 
 
 def main(argv=None):
@@ -306,8 +170,6 @@ def main(argv=None):
         else SPNTA.load_jlso(args.jlso_file, args.par_file)
     )
 
-    save_spnta_attrs(spnta, args)
-
     nwalkers = spnta.ndim * 5
     p0 = np.array(
         [spnta.prior_transform(cube) for cube in np.random.rand(nwalkers, spnta.ndim)]
@@ -323,38 +185,17 @@ def main(argv=None):
     )
     sampler.run_mcmc(p0, args.nsteps, progress=True, progress_kwargs={"mininterval": 1})
     samples_raw = sampler.get_chain(flat=True, discard=args.burnin, thin=args.thin)
-    samples = spnta.rescale_samples(samples_raw)
 
-    with open(f"{args.outdir}/samples_raw.npy", "wb") as f:
-        np.save(f, samples_raw)
-    with open(f"{args.outdir}/samples.npy", "wb") as f:
-        np.save(f, samples)
-
-    param_uncertainties = np.std(samples_raw, axis=0)
-
-    params_maxpost = sampler.chain[
-        *np.unravel_index(
-            np.argmax(sampler.get_log_prob().T), sampler.get_log_prob().T.shape
-        ),
-        :,
-    ]
-    np.savetxt(f"{args.outdir}/params_maxpost.txt", params_maxpost)
-    save_new_parfile(
-        spnta,
-        params_maxpost,
-        param_uncertainties,
-        f"{args.outdir}/{spnta.model.pulsar_name}.maxpost.par",
+    spnta.save_results(
+        args.outdir,
+        samples_raw,
+        {
+            "sampler": "emcee",
+            "nwalkers": nwalkers,
+            "nsteps": args.nsteps,
+            "burnin": args.burnin,
+            "thin": args.thin,
+            "vectorized": True,
+        },
+        args.truth,
     )
-
-    params_median = np.median(samples_raw, axis=0)
-    np.savetxt(f"{args.outdir}/params_median.txt", params_median)
-    save_new_parfile(
-        spnta,
-        params_median,
-        param_uncertainties,
-        f"{args.outdir}/{spnta.model.pulsar_name}.median.par",
-    )
-
-    save_resids(spnta, params_median, args.outdir)
-
-    np.savetxt(f"{args.outdir}/param_default_values.txt", spnta.default_params)